Face Recogntion Convulsional Neural Networks
Here we use the convulsion neural networks to identify the person in an image and a video. I have taken the Malayalam actor Nivin Pauly for example.
The dataset image contains images to be trained of Nivin Pauly.We can specify whether to use the method of CNN or HOG.
The working of CNN can be viewed here:https://medium.com/@ageitgey/machine-learning-is-fun-part-4-modern-face-recognition-with-deep-learning-c3cffc121d78
The trained network that we use is developed by Davis King.The description of that is given:http://blog.dlib.net/2017/02/high-quality-face-recognition-with-deep.html
The face_recognition  library, created by Adam Geitgey, wraps around dlib’s facial recognition functionality, making it easier to work with.
The encode_faces.py converts the dataset images into a series of codes that is found out using the network dlib,into the file encodings.pickle
This file is then used always to check the validity of test images and videos.
For thorough understanding,I request you to follow Adrian Rosebrock on his website:https://www.pyimagesearch.com/2018/06/18/face-recognition-with-opencv-python-and-deep-learning/
This is an adaptation of his tutorial code.